+++ plugins/group-sort-strapi-plugin/admin/src/components/GroupAndArrangeContextProvider.tsx
@@ -2,16 +2,17 @@ import { useLocalStorage } from "react-use";
 import { LocalConfig, LocalSettings, OrderFieldConfiguration, Settings } from "../../../shared/settings";
 import { LOCAL_SETTINGS_KEY } from "../../../shared/constants";
 import { useParams } from "react-router-dom";
-import { createContext, useEffect, useState } from "react";
+import { createContext, useCallback, useEffect, useState } from "react";
 import useCollectionTypes from "../hooks/useCollectionTypes";
 import useLocalConfig from "../hooks/useLocalConfig";
 import useGroupData from "../hooks/useGroupData";
 import useAttributeData from "../hooks/useAttributeData";
-import { Attribute } from '@strapi/types/dist/schema';
 import { GroupResult, GroupResultMeta } from "../../../shared/contracts";
+import { GridDirection } from "../../../shared/types";
 import useGroupNames from "../hooks/useGroupNames";
 import { ContentTypeSchema } from "@strapi/types/dist/struct";
 import useSettings from "../hooks/useSettings";
+import { OrderAttribute } from "../types";
 
 export const GroupAndArrangeContext = createContext<GroupAndArrangeContextValue & GroupAndArrangeContextSetters>(undefined!);
 
@@ -26,13 +27,10 @@ export interface GroupAndArrangeContextValue {
   chosenMediaField: string | null;
   chosenTitleField: string | null;
   chosenSubtitleField: string | null;
-  chosenDirection: 'horizontal' | 'vertical' | null;
+  chosenDirection: GridDirection;
   mediaAttributeNames: string[];
   titleAttributeNames: string[];
-  currentAttribute: (Attribute.AnyAttribute & {
-    isOrder: boolean;
-    isOrder2d: boolean;
-  }) | null;
+  currentAttribute: OrderAttribute | null;
   currentCollectionType: ContentTypeSchema | null;
   currentFieldSettings: OrderFieldConfiguration | null;
   groupData: GroupResult | null;
@@ -44,22 +42,27 @@ export interface GroupAndArrangeContextValue {
 export interface GroupAndArrangeContextSetters {
   setLocalSettings: (newConfig: LocalSettings) => void;
   setLocalConfig: (config: any) => void;
-  setChosenDirection: (direction: 'horizontal' | 'vertical' | null) => void;
+  setChosenDirection: (direction: GridDirection) => void;
+  // triggerUpdate is a function that can be called to trigger a re-fetch all of the data from the API
+  triggerUpdate: () => void;
 }
 
 /**
- * Provider for the GroupAndArrangeContext, holds pretty much all the state for the plugin
+ * Provider for the GroupAndArrangeContext, holds pretty much all the state for the plugin, also responsible for fetching data from the API
  */
 export const GroupAndArrangeContextProvider = ({ children }: { children: React.ReactNode }) => {
+  const [updateCounter, setUpdateCounter] = useState(0);
+  const triggerUpdate = useCallback(() => setUpdateCounter((prev) => prev + 1), [setUpdateCounter]);
+
   const [localSettings, setLocalSettings] = useLocalStorage<LocalSettings>(LOCAL_SETTINGS_KEY, {
     configs: {},
   });
   const { uid: contentTypeUid, groupField, groupName } = useParams<{ uid: string, groupField: string, groupName: string }>();
-  const { collectionTypes, isLoading: isLoadingCollectionTypes } = useCollectionTypes();
+  const { collectionTypes, isLoading: isLoadingCollectionTypes } = useCollectionTypes({ updateCounter });
   const [localConfig, setLocalConfig] = useLocalConfig({ contentTypeUid, groupField, groupName, localSettings, setLocalSettings });
-  const { groupData, isLoading: isLoadingGroupData } = useGroupData({ contentTypeUid, groupField, groupName });
-  const { groupNames, isLoading: isFetchingGroupNames } = useGroupNames({ contentTypeUid });
-  const { settings, isLoading: isFetchingSettings } = useSettings();
+  const { groupData, isLoading: isLoadingGroupData } = useGroupData({ contentTypeUid, groupField, groupName, updateCounter });
+  const { groupNames, isLoading: isFetchingGroupNames } = useGroupNames({ contentTypeUid, updateCounter });
+  const { settings, isLoading: isFetchingSettings } = useSettings({ updateCounter });
 
   const { chosenMediaField, chosenTitleField, chosenSubtitleField, mediaAttributeNames, titleAttributeNames, currentAttribute, currentCollectionType, currentFieldSettings } = useAttributeData({
     contentTypeUid,
@@ -99,7 +102,8 @@ export const GroupAndArrangeContextProvider = ({ children }: { children: React.R
     globalSettings: settings || null,
     setLocalSettings,
     setLocalConfig,
-    setChosenDirection
+    setChosenDirection,
+    triggerUpdate
   };
 
   return (
