+++ plugins/group-sort-strapi-plugin/admin/src/pages/ArrangePage.tsx
@@ -1,7 +1,7 @@
 import { Box, Button, Flex, Grid, Main, StrapiTheme } from '@strapi/design-system';
 import { useIntl } from 'react-intl';
 import { useTranslation } from '../hooks/useTranslation';
-import GridLayout, { WidthProvider } from "react-grid-layout";
+import GridLayout from "react-grid-layout";
 
 import styled from 'styled-components';
 import { Layouts, Page, useFetchClient, useNotification } from '@strapi/strapi/admin';
@@ -9,17 +9,13 @@ import { useContext, useEffect, useState } from 'react';
 
 import { LeftMenu } from '../components/LeftMenu';
 import { Check } from '@strapi/icons';
-import withReactGridStyles from './ArrangePage.Styles';
-import { get, isEqual, differenceWith } from 'lodash';
-import { DndContext, DragEndEvent, KeyboardSensor, PointerSensor, closestCenter, useSensor, useSensors } from '@dnd-kit/core';
-import { SortableContext, arrayMove, rectSortingStrategy, sortableKeyboardCoordinates } from '@dnd-kit/sortable';
-import { SortableItem } from '../components/ArrangePage/SortableItem';
+import { get } from 'lodash';
 import { UserSettings } from '../components/ArrangePage/UserSettings';
 import { GroupAndArrangeContext } from '../components/GroupAndArrangeContextProvider';
 import { useBlocker } from 'react-router-dom';
-
-
-const StyledResponsiveGridLayout = withReactGridStyles(WidthProvider(GridLayout));
+import { ItemsDictionary } from '../types'
+import Order1dComponent from '../components/ArrangePage/Order1dComponent';
+import Order2dComponent from '../components/ArrangePage/Order2dComponent';
 
 const StyledUserSettings = styled(UserSettings)`
   align-items: flex-start;
@@ -43,17 +39,11 @@ const ArrangePage = () => {
     groupName,
     groupData,
     currentAttribute,
-    currentFieldSettings,
     mediaAttributeNames,
     titleAttributeNames,
-    chosenMediaField,
-    chosenTitleField,
-    chosenSubtitleField,
-    chosenDirection,
-    localConfig,
     currentCollectionType,
-    collectionTypes,
-    isLoading
+    isLoading: isLoadingExt,
+    triggerUpdate
   } = useContext(GroupAndArrangeContext);
   if (!contentTypeUid || !groupField || !groupName) {
     return <Page.Error />;
@@ -65,10 +55,13 @@ const ArrangePage = () => {
     setIsModified(false);
     setIsSaving(false);
   }, [contentTypeUid, groupField, groupName]);
+  
+  const [layout1d, setLayout1d] = useState([] as string[]);
+  const [layout2d, setLayout2d] = useState([] as GridLayout.Layout[]);
 
-  const [itemsDictionary, setItemsDictionary] = useState({} as Record<string, any>);
+  const [itemsDictionary, setItemsDictionary] = useState({} as ItemsDictionary);
   useEffect(() => {
-    const itemsDictionary = (groupData?.items || []).reduce((acc: Record<string, any>, item) => {
+    const itemsDictionary = (groupData?.items || []).reduce((acc: ItemsDictionary, item) => {
       acc[item.documentId] = {
         ...item,
         thumbnailUrisByMediaFields: mediaAttributeNames.reduce((mediaAcc: Record<string, string>, mediaAttributeName) => {
@@ -91,115 +84,14 @@ const ArrangePage = () => {
     }, {});
 
     setItemsDictionary(itemsDictionary);
-    if (currentAttribute?.isOrder) {
-      // Prepare sortables for 1-dimensional list.
-      // Items that does not have order yet will be placed at the end of the list.
-      const sortables: {order: number, documentId: string}[] = groupData?.items.reduce((acc: {order: number, documentId: string}[], item) => {
-        const order = item[groupField] || acc.length - 1;
-        acc.push({ order, documentId: item.documentId });
-        return acc;
-      }, []) || [];
-      sortables.sort((a, b) => a.order - b.order);
-      setSortables(sortables.map(x => x.documentId));
-    }
-    if (currentAttribute?.isOrder2d && currentFieldSettings) {
-      // Prepare layout2d for 2-dimensional grid.
-      // Items that does not have order yet will be placed at the end of the grid starting from the last available coordinate.
-      const sortables: {order: GridLayout.Layout | null, documentId: string}[] = 
-        groupData?.items
-          .reduce((acc: {order: GridLayout.Layout | null, documentId: string}[], item) => {
-            const order = item[groupField];
-            if (typeof order !== 'object' && order !== null && order !== undefined) {
-              return null;
-            }
-            acc.push({ order: order || null, documentId: item.documentId });
-            return acc;
-          }, [])
-          ?.filter((x: any) => x)
-          .map((x: { order: GridLayout.Layout | null; documentId: string; }) => {
-            // attach documetId since it is not part of the order object
-            return { order: x.order && {...x.order, i: x.documentId} as GridLayout.Layout, documentId: x.documentId }
-          }) || [];
-      
-      const resultingSortables = sortables?.filter(x => x.order && x.order.x !== undefined && x.order.y !== undefined && x.order.w !== undefined && x.order.h !== undefined).map(x => x.order as GridLayout.Layout);
-      // sortables except resultingSortables
-      const idsMissingOrder = sortables?.filter(x => !resultingSortables.some(y => y.i === x.documentId)).map(x => x.documentId);
-      if(idsMissingOrder?.length > 0) {
-        let lastPositionY = 0;
-        resultingSortables.forEach(i => {
-          lastPositionY = Math.max(lastPositionY, i.y + i.h);
-        });
-        for(let i = 0; i < idsMissingOrder.length; i++)
-        {
-          const id = idsMissingOrder[i];
-          const y = Math.floor(i / currentFieldSettings.columnsNumber);
+  }, [groupData]);
 
-          const layout: GridLayout.Layout = {
-            i: id,
-            x: 0,
-            y: lastPositionY + y,
-            w: 1,
-            h: 1
-          }
-          resultingSortables.push(layout);
-        }
-      }
-      setLayout2d(resultingSortables.map(x => ({
-        ...x,
-        // these are needed for checking whether changes were made in handleLayout2dChange:
-        moved: false,
-        static: false,
-        isBounded: undefined,
-        isDraggable: undefined,
-        isResizable: undefined,
-        maxH: undefined,
-        maxW: undefined,
-        minH: undefined,
-        minW: undefined,
-        resizeHandles: undefined
-      })));
-    }
-
-  }, [groupData, collectionTypes, currentAttribute, currentFieldSettings]);
-  const [sortables, setSortables] = useState([] as string[]);
-  const [layout2d, setLayout2d] = useState([] as GridLayout.Layout[]);
-
-  const sensors = useSensors(
-    useSensor(PointerSensor),
-    useSensor(KeyboardSensor, {
-      coordinateGetter: sortableKeyboardCoordinates,
-    })
-  );
-
-  function handleDragEnd(event: DragEndEvent): void {
-    const { active, over } = event;
-
-    if (active.id !== over?.id) {
-      setSortables((items) => {
-        const oldIndex = items.indexOf(active.id as string);
-        const newIndex = items.indexOf(over?.id as string);
-
-        setIsModified(true);
-        return arrayMove(items, oldIndex, newIndex);
-      });
-    }
-  }
-  function handleLayout2dChange(layout: GridLayout.Layout[]): void {
-    if(!isModified) {
-      const wasChanged = differenceWith(layout, layout2d, isEqual).length > 0;
-      if (!wasChanged) {
-        return;
-      }
-      setIsModified(true);
-    }
-    setLayout2d(layout.filter(x => x && x.i && x.x !== undefined));
-  }
 
   async function handleSave(): Promise<void> {
     setIsSaving(true);
 
     if (currentAttribute?.isOrder) {
-      const currentValues = sortables.map((documentId, index) => ({ documentId, index }));
+      const currentValues = layout1d.map((documentId, index) => ({ documentId, index }));
       for (const { documentId, index } of currentValues) {
         await fetchClient.put(`/content-manager/collection-types/${currentCollectionType?.uid}/${documentId}`, {
           [groupField!]: index
@@ -208,20 +100,21 @@ const ArrangePage = () => {
     }
     if (currentAttribute?.isOrder2d) {
       for (const item of layout2d) {
-        const stringified = {
+        const value = {
           x: item.x,
           y: item.y,
           w: item.w,
           h: item.h
         };
         await fetchClient.put(`/content-manager/collection-types/${currentCollectionType?.uid}/${item.i}`, {
-          [groupField!]: stringified
+          [groupField!]: value
         });
       }
     }
 
     setIsModified(false);
     setIsSaving(false);
+    triggerUpdate();
 
     toggleNotification({
       type: 'success',
@@ -244,9 +137,11 @@ const ArrangePage = () => {
     return false;
   });
 
-  if (isLoading || !currentFieldSettings || isSaving) {
-    return <Page.Loading />;
-  }
+  // if (isLoading || !currentFieldSettings || isSaving) {
+  //   return <Page.Loading />;
+  // }
+
+  const isLoading = isLoadingExt || isSaving;
 
   return (
     <Layouts.Root sideNav={<LeftMenu />}>
@@ -270,61 +165,28 @@ const ArrangePage = () => {
               })}
             </Button>
           </Flex>} />
-      <Main style={{ marginTop: "-24px" }}>
-        <MainBox>
-          <Flex direction="column" alignItems="stretch" gap={4}>
-            <StyledUserSettings />
-            {currentAttribute?.isOrder &&
-              <Grid.Root gap={4} gridCols={currentFieldSettings.columnsNumber} flex={1}>
-                <DndContext
-                  sensors={sensors}
-                  collisionDetection={closestCenter}
-                  onDragEnd={handleDragEnd}
-                >
-                  <SortableContext
-                    items={sortables}
-                    strategy={rectSortingStrategy}
-                  >
-                    {sortables.map(x => itemsDictionary[x]).filter(x => x).map(i =>
-                      <Grid.Item key={i.documentId} col={1} m={1}>
-                        <SortableItem
-                          id={i.documentId}
-                          title={chosenTitleField && i.titlesByTitleFields[chosenTitleField]}
-                          subtitle={chosenSubtitleField && i.titlesByTitleFields[chosenSubtitleField]}
-                          thumbnailUri={chosenMediaField && i.thumbnailUrisByMediaFields[chosenMediaField]}
-                          resizable={false} />
-                      </Grid.Item>
-                    )}
-                  </SortableContext>
-                </DndContext>
-              </Grid.Root>
-            }
-            {currentAttribute?.isOrder2d &&
-              <Box position="relative" width="100%">
-                <StyledResponsiveGridLayout
-                  className="layout"
-                  layout={layout2d}
-                  onLayoutChange={handleLayout2dChange}
-                  cols={currentFieldSettings.columnsNumber}
-                  rowHeight={localConfig?.rowHeight || 30}
-                  compactType={chosenDirection || null}
-                >
-                  {layout2d.map(x => itemsDictionary[x.i]).filter(x => x).map(i =>
-                    <Grid.Item key={i.documentId} col={1} m={1}>
-                      <SortableItem
-                        id={i.documentId}
-                        title={chosenTitleField && i.titlesByTitleFields[chosenTitleField]}
-                        subtitle={chosenSubtitleField && i.titlesByTitleFields[chosenSubtitleField]}
-                        thumbnailUri={chosenMediaField && i.thumbnailUrisByMediaFields[chosenMediaField]}
-                        resizable={true} />
-                    </Grid.Item>
-                  )}
-                </StyledResponsiveGridLayout>
-              </Box>
-            }
-          </Flex>
-        </MainBox>
-      </Main>
+      {isLoading && <Page.Loading />}
+      {!isLoading &&
+        <Main style={{ marginTop: "-24px" }}>
+          <MainBox>
+            <Flex direction="column" alignItems="stretch" gap={4}>
+              <StyledUserSettings />
+              {currentAttribute?.isOrder &&
+                <Order1dComponent
+                  itemsDictionary={itemsDictionary}
+                  setIsModified={setIsModified}
+                  setLayout1d={setLayout1d} />
+              }
+              {currentAttribute?.isOrder2d &&
+                <Order2dComponent
+                  itemsDictionary={itemsDictionary}
+                  isModified={isModified}
+                  setIsModified={setIsModified}
+                  setLayout2d={setLayout2d} />
+              }
+            </Flex>
+          </MainBox>
+        </Main>}
     </Layouts.Root>
   );
 };
