+++ plugins/group-sort-strapi-plugin/shared/types.ts
@@ -1,0 +1,1 @@
+export type GridDirection = 'horizontal' | 'vertical' | null;
